+++ schronisko/src/components/DogsList/SingleDogViewForClients.jsx
@@ -8,9 +8,7 @@ import { GetDogPhotoForList } from "../DogProfile/GetDogPhotoForList";
 import Container from "@mui/material/Container";
 import { Link } from "react-router-dom";
 import DescriptionWithoutEdit from "./DescriptionWithoutEdit.jsx";
-import { useNavigate } from "react-router-dom";
-import AddIcon from "@mui/icons-material/Add";
-import { Button } from "@mui/material";
+import AskButton from "./AskButton.jsx";
 
 const Item = styled(Paper)(({ theme }) => ({
   backgroundColor: "#f6f7fa",
@@ -22,8 +20,6 @@ const Item = styled(Paper)(({ theme }) => ({
 }));
 
 const SingleDogViewForClients = ({ id }) => {
-  const navigate = useNavigate();
-
   return (
     <Container
       sx={{
@@ -49,9 +45,9 @@ const SingleDogViewForClients = ({ id }) => {
           </Grid>
         </Grid>
 
-        <Grid item xs={16} sm={4} md={2}>
-          <Grid container spacing={0} columns={2}>
-            <Grid item xs={1} sm={2} md={2}>
+        <Grid item xs={16} sm={4} md={3}>
+          <Grid container spacing={0} columns={3}>
+            <Grid item xs={1} sm={3} md={3}>
               <Item
                 sx={{
                   padding: "0",
@@ -61,7 +57,7 @@ const SingleDogViewForClients = ({ id }) => {
                 <IconNameBreed id={id} />
               </Item>
             </Grid>
-            <Grid item xs={1} sm={2} md={2}>
+            <Grid item xs={1} sm={3} md={3}>
               <Item
                 sx={{
                   padding: "0",
@@ -74,9 +70,9 @@ const SingleDogViewForClients = ({ id }) => {
           </Grid>
         </Grid>
 
-        <Grid item xs={16} sm={4} md={6}>
-          <Grid container spacing={0} columns={6}>
-            <Grid item xs={6} sm={6} md={6}>
+        <Grid item xs={16} sm={4} md={4}>
+          <Grid container spacing={0} columns={4}>
+            <Grid item xs={4} sm={4} md={4}>
               <Item
                 sx={{
                   padding: "0",
@@ -88,24 +84,16 @@ const SingleDogViewForClients = ({ id }) => {
             </Grid>
           </Grid>
         </Grid>
-        <Grid item xs={16} sm={4} md={6}>
-          <Grid container spacing={0} columns={6}>
-            <Grid item xs={6} sm={6} md={6}>
+        <Grid item xs={16} sm={4} md={3}>
+          <Grid container spacing={0} columns={3}>
+            <Grid item xs={3} sm={3} md={3}>
               <Item
                 sx={{
                   padding: "0",
                 }}
                 elevation={0}
               >
-                <Button
-                  variant="contained"
-                  startIcon={<AddIcon />}
-                  size="large"
-                  onClick={() => navigate("/adopt")}
-                  color="primary.light"
-                >
-                  Adoptuj
-                </Button>
+                {/* <AskButton /> */}
               </Item>
             </Grid>
           </Grid>
